requirements.txt
fastapi
sqlalchemy
strawberry-graphql
uvicorn[standard]
pytest
httpx
